[Refactor] Allow shared lib use via compile time flag.
By default, the build now uses a shared library approach
to load the QRCode library. The user needs to build the
*.DLL or *.so file before using the project. I'm doing it
this way so that the JITx repl can properly load the project.

The user can set the `COMPILE-STATIC` flag in the stanza build
and force the code to use a static library to load the C
functions for the QRcode implementation at build time.

Note that this is supposedly temporary and that a future stanza version
will make this unnecessary.

This adds content to the Makefile to build the *.so file and a
bat script to build the DLL file on windows assuming mingw64
exists. This is admittedly very hacky right now.